Refactor and linting
