+++ lib/client.js
@@ -1,10 +1,10 @@
-require('abort-controller/polyfill');
 const queryString = require('query-string');
 const { authFromCapability, authFromWellKnown } = require('./smart');
 const HttpClient = require('./http-client');
 const ReferenceResolver = require('./reference-resolver');
 const Pagination = require('./pagination');
 const { createQueryString } = require('./utils');
+const { FetchQueue } = require('./fetch-queue');
 const { deprecatePaginationArgs, deprecateHeaders } = require('./deprecations');
 
 /**
@@ -177,8 +177,10 @@ class Client {
   }
 
   /**
-   * Obtain the SMART OAuth URLs from the Capability Statement
-   * http://docs.smarthealthit.org/authorization/conformance-statement/
+   * Obtain the SMART OAuth URLs from the Capability Statement, or
+   * any of the .well-known addresses.
+   *
+   * See: http://docs.smarthealthit.org/authorization/conformance-statement/
    *
    * @async
    *
@@ -198,46 +200,14 @@ class Client {
    * @param {Object} [params.options.headers] - Optional headers to add to the
    *   request
    *
-   * @return {Promise<Object>} contains the following SMART URIs: authorizeUrl,
+   * @return {Promise<Object>} contains the following SMART URL: authorizeUrl,
    *   tokenUrl, registerUrl, manageUrl
    */
   async smartAuthMetadata({ headers, options = {} } = {}) {
-    class FetchJob {
-      constructor() {
-        this.controller = new AbortController();
-        this.resolving = false;
-      }
-
-      addSignalOption(options) {
-        return {
-          signal: this.controller.signal,
-          ...options,
-        };
-      }
-    }
-
-    class FetchQueue {
-      constructor() {
-        this.jobs = [];
-        this.numJobs = 0;
-      }
-
-      buildJob() {
-        const job = new FetchJob();
-        this.numJobs = this.jobs.push(job);
-        return job;
-      }
-
-      safeAbortOthers(currentJob) {
-        for (let a = 0, l = this.numJobs; a < l; ++a) {
-          if (currentJob === this.jobs[a]) { continue; }
-          this.jobs[a].controller.abort();
-        }
-      }
-    }
+    const fetchOptions = deprecateHeaders(options, headers);
+    const normalizedBaseUrl = this.baseUrl.replace(/\/*$/, '/');
 
     const queue = new FetchQueue();
-
     const metadataJob = queue.buildJob();
     const wellknownSmartJob = queue.buildJob();
     const wellknownOidcJob = queue.buildJob();
@@ -251,23 +221,27 @@ class Client {
         }
       }
 
-      const wellKnownSmartUrl = `${this.baseUrl.replace(/\/*$/, '/')}.well-known/smart-configuration`;
-      this.httpClient.request('GET', wellKnownSmartUrl, wellknownSmartJob.addSignalOption(options))
+      const wellKnownSmartUrl = `${normalizedBaseUrl}.well-known/smart-configuration`;
+      this.httpClient.request('GET',
+        wellKnownSmartUrl,
+        wellknownSmartJob.addSignalOption(fetchOptions))
         .then((r) => {
           queue.safeAbortOthers(wellknownSmartJob);
           return resolve(authFromWellKnown(r));
         })
         .catch((e) => handleError(e));
 
-      this.capabilityStatement(metadataJob.addSignalOption({ options: deprecateHeaders(options, headers) }))
+      this.capabilityStatement(metadataJob.addSignalOption(fetchOptions))
         .then((r) => {
           queue.safeAbortOthers(metadataJob);
           return resolve(authFromCapability(r));
         })
         .catch((e) => handleError(e));
 
-      const wellKnownOidcUrl = `${this.baseUrl.replace(/\/*$/, '/')}.well-known/openid-configuration`;
-      this.httpClient.request('GET', wellKnownOidcUrl, wellknownOidcJob.addSignalOption(options))
+      const wellKnownOidcUrl = `${normalizedBaseUrl}.well-known/openid-configuration`;
+      this.httpClient.request('GET',
+        wellKnownOidcUrl,
+        wellknownOidcJob.addSignalOption(fetchOptions))
         .then((r) => {
           queue.safeAbortOthers(wellknownOidcJob);
           return resolve(authFromWellKnown(r));
