+++ lib/smart.js
@@ -50,6 +50,7 @@ function authFromCapability(capabilityStatement) {
   }
 }
 
+/* eslint-disable camelcase */
 function authFromWellKnown(wellKnown) {
   const {
     authorization_endpoint,
